wssmail: Initial commit.
This is the initial commit of the wssmail frontend.
A compatible backend (of which there is currently only 1)
is required to get a full working wssmail system.

Currently, wssmail is minimally functional at what a mail
client should be able to do, at minimum (and even quite a
bit more than that, in some areas). There is still a lot
of stuff to be done, but this captures a "base working version".