FETCH/SEEN/UNSEEN: Update message listing/counts automatically.
When a message is viewed and implicitly marked as seen, the
message listing table is now updated so that this displays.

A more major change is that the folder counts (just the unread count
in this case) now update when messages are marked as read/unread,
either implicitly by viewing them or explicitly by marking them.