FETCH: Use iframe to isolate HTML emails.
Previously, when viewing HTML emails, their
content was just injected into the page, which
was very much the wrong way to do this. Rather
than trying to sanitize the HTML though, we
opt to display it in an iframe, which takes
care of all of that for us.

The webpage title now also updates in response
to changes in the unread count in a folder.