FETCH: Fix HTML message display height.
When viewing a message with the preview pane disabled,
rather than the default (tiny) default frame height,
show as much of the frame as possible with the
available viewport height. There isn't a good way
to do this, so we just arbitrarily subtract 200px
from 100%, which should make it fill the majority
of the page.