FETCHLIST: Prevent right and middle-clicking.
I know this is evil, but there is a good reason for this.
Beta testers have apparently been trying to right-click
message links to open them in a new tab, but are puzzled
when the folder just opens again. This is because it's
completely separate from the parent, and you just can't
do this, since the parent needs to spawn the child.
Disabling the preview pane and left-clicking achieves
the same result, although it won't provide a persistent
URL to the message (since there isn't one, currently).
This is by design as that would require setting up
a separate TCP and IMAP connection just to load a
message requested in this manner.