FETCH: Fix messaging pseudo formatting.
This fixes overly broad regex matches for
detecting *bold*, /italics/, and _underline_
in plain text messages.

We now require a space before the special
pseudo formatting character, to avoid matching
strings such as URLs with slashes anywhere
after the protocol, for instance. This also
fixes the case of these characters being used
casually in a message, when not intended as
formatting hints.