LIST: Fix first folder not showing details.
Compute showtotal before the first folder is
added to the folder list so that it displays
correctly.

This also adds some WIP code for making subfolder
lists collapsible, but there is an issue where
nested folders only take up the first column
of their parent (with display: table). The HTML
structure is correct, but the styling here is
not right. Therefore, this is on hold until
that is resolved. We cannot use a table with
colspan because we need a hierarchical list
for being able to set a child container to
display: none when collapsed, for example.