FETCHLIST: Allow all messages in folder to be selected.
Currently, it is only possible to select messages that
are visible on the current page (e.g. by selecting the
first message, holding SHIFT, and selecting the last).
Because the page size is limited, it was therefore
not possible to perform an operation on all messages in
a mailbox in one go.

We now support performing an operation on all messages
in the current mailbox if there are messages that are
not visible on the current page. The user is currently
prompted for the scope of the selected messages (entire
mailbox or only current folder). We can then reference
all messages using 1:* instead of enumerating UIDs,
which is clearly wholly impractical for large mailboxes.

As an example, this now makes it much easier to do things
like "empty all messages in the trash", since all messages
can be marked as deleted, and then expunged easily.
Previously, this required repeating the first step of
the procedure for every page of messages.