SELECT: Fix auto-unmark folder when selected.
Mailboxes that were "Marked" are internally no
longer marked when selected, to reflect the
behavior of most servers. However, this was
not visually updating immediately, only when
a message was selected or some other operation
was performed that caused the folder pane to
update. The folder pane is now updated immediately
whenever a message is implicitly no longer marked.