FETCH: Fix remote content blocking.
In some cases, the logic to add a CSP meta tag
prohibiting remote content did not work properly,
adding the tag to the body instead of the head tag.
Rather than trying to inject it into the markup
before the new DOM is built, we now use DOM manipulation
to ensure it's properly added to the head. This more
robustly ensures that the CSP meta tag is obeyed,
preventing remote content from inadvertently loading
when not desired.