FETCH/FETCHLIST: Improve message viewing/downloading UI.
We now display sent/received times without the date only if
the timestamp is today, as opposed to within the past 24 hours.
This is more in line with user expectations and behavior of
other mail clients, like Thunderbird.

When the raw message source is downloaded (either to display
in browser or download as a .eml file), the status bar indicates
a download is in progress to the user, since this could take
a long time on dial-up connections and users might be wondering
if anything is happening or if something got stuck.