FETCHLIST: Set preview pane height to 0 on each response.
The preview pane height is currently set to 0 whenever it is toggled
on or off, but this doesn't cover all cases where this is needed.
Also set the height to 0 explicitly whenever we get a FETCHLIST
response, to ensure that the stale height from a previous operation
doesn't inadvertently push the preview pane off the bottom of the
display and add an unwanted scrollbar to the page. The preview pane
height is still set as needed when a message is selected, if the
preview pane is enabled.

WSSMAIL-10 #close