wssmail: Add click-and-drag to move messages.
Allow messages to be selected more easily by clicking the subject
and using CTRL and SHIFT modifiers for selection, and allow
dragging a selection to another folder to move messages.
This is an easier alternative to using the existing copy/move-to
folder dropdown input and clicking "Move", and mirrors functionality
typically available in desktop mail clients.