viewer: Fix bugs with quote rendering.
A bug existed where if the quote depth changed by more than 1 at a time,
only one quote bar was added or removed in the viewer, which resulted
in an inaccurate quoting hierarchy being displayed. We now take care to
add or remove the necessary number of quoting bars (divs) when the quoting
depth changes.