smtp.php: Manually generate the Message-ID.
PHPMailer defaults to a Message-ID that contains the HTTP hostname,
rather than any hostname used in the SMTP transaction or the email
message. This is problematic if it is an IP address, and even worse,
this can leak sensitive information. So now, we manually generate
our own Message-ID, consisting of a UUID for the user portion and
the domain used in the From address for the domain portion.