FETCHLIST: Increase padding of subject column.
Add some padding to the left and right in the subject column.
This increases the surface area available for click+drag
operations, making it easier to move messages to a folder,
particularly if the subject is long and thus the text/link
takes up almost all the room, which previously left very
little to no room for selection for dragging.