webmail.js: Clear status message upon page interaction.
If a user interacts with the application (clicks on a message,
selects a folder, or copies/moves a message), clear any previous
status message to avoid confusion with if the message was
intended for a previous action or the most recent event.

As a result, as soon as a raw message finishes downloading,
we now clear the status message that told the user to wait for
the message to download.