FETCH: Display message if plain text part missing or empty.
If the message does not contain a plain text component, show a message
to the user to that effect, suggesting viewing the HTML version
instead. Otherwise, nothing was displayed and it wasn't clear if
something was broken.

Similarly, if the message contains a plain text, but it's empty,
we would crash calling displayFormatFlowed() on undefined; instead,
show a similar message indicating the part is empty.