FETCHLIST: Fix next page navigation.
pageNumber was being used as a String, rather than as a number,
leading to improper arithmetic operations that resulted in
jumping to the last page when clicking "Next >>", since the
conditional would erroneously evaluate to true. Explicitly
convert parsed in values for page number to a number so that
the math works correctly.